Mute ctags "ignoring null tag in" warnings for JavaScript files
This warning being generated a lot is a known issue for JavaScript files
and seems to be related to object destructuring.

The new `ctags` function takes precedence over the `ctags` executable,
runs the actual `ctags` binary but filters its stderr output for the
specific warnings being thrown for JavaScript files.