Allow search using ripgrep while ignoring VCS ignored files
Ripgrep is configured (in `ripgrep.conf`) to *do* search in files
ignored by the version control system. This also makes `:Rg` search in
those files.

However sometimes I want to specifically *not* search in ignored files,
so let's add a `:Rgi` command that does that, similar to the `rgi` shell
alias.