Fix PHP files with PHP-CS-Fixer using custom fixer
The built-in ALE fixer using PHP-CS-Fixer works with temporary files.
This does not work when using a Docker development environment where the
fixer is running in a tailored container.

This custom fixer callback creates a temporary copy of the current file
in the working directory, runs `php-cs-fixer fix` on the file using a
relative path, then feeds the result back to the buffer, deleting the
copy.