Open ALE location list only on buffer save
It seems opening the location list switches modes (or a similar
operation) that makes it impossible to complete UltiSnips snippets and
cycle to next/previous variables. Since this is an essential part of my
workflow, I change ALE to only open the loclist when saving, but keep
linting active when returning to normal mode.